Refs #2379: Splitting out template cruft out so it is easier to containerise
